backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
PyPDF2==3.0.1
pdfplumber==0.10.3
tabula-py==2.9.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.26.2
httpx==0.25.2
beautifulsoup4==4.12.2
google-generativeai==0.8.3
reportlab==4.0.8
python-docx==1.1.0
jinja2==3.1.2
python-dotenv==1.0.0
scikit-learn==1.3.2
aiofiles==23.2.1
sqlalchemy==2.0.23
openai==1.6.1
joblib==1.3.2
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.1.0
